Mes 3644 restructure contact details page (#817)
* Mes 3644 restructure folders for contact details page
