+++ src/pages/communication-page/cat-b/components/new-email/__tests__/new-email.spec.ts
@@ -1,7 +1,7 @@
 import { ComponentFixture, async, TestBed } from '@angular/core/testing';
 import { IonicModule } from 'ionic-angular';
 import { By } from '@angular/platform-browser';
-import * as welshTranslations from '../../../../../assets/i18n/cy.json';
+import * as welshTranslations from '../../../../../../assets/i18n/cy.json';
 import { TranslateModule, TranslateService, TranslateLoader } from 'ng2-translate';
 import { FormGroup, FormControl } from '@angular/forms';
 import { Http } from '@angular/http';
