+++ src/pages/communication-page/cat-b/components/privacy-notice/__tests__/privacy-notice.spec.ts
@@ -2,7 +2,7 @@ import { ComponentFixture, async, TestBed } from '@angular/core/testing';
 import { IonicModule, Config } from 'ionic-angular';
 import { PrivacyNoticeComponent } from '../privacy-notice';
 import { TranslateService, TranslateModule } from 'ng2-translate';
-import { translateServiceMock } from '../../../../../shared/__mocks__/translate';
+import { translateServiceMock } from '../../../../../../shared/__mocks__/translate';
 import { ConfigMock } from 'ionic-mocks';
 
 describe('PrivacyNoticeComponent', () => {
