+++ src/pages/communication-page/cat-b/components/privacy-notice/privacy-notice.scss
