+++ src/pages/communication-page/cat-b/components/provided-email/__tests__/provided-email.spec.ts
@@ -4,8 +4,8 @@ import { FormGroup, FormControl } from '@angular/forms';
 import { IonicModule } from 'ionic-angular';
 import { By } from '@angular/platform-browser';
 import { TranslateService, TranslateModule, TranslateLoader } from 'ng2-translate';
-import * as welshTranslations from '../../../../../assets/i18n/cy.json';
-import { createTranslateLoader } from '../../../../../app/app.module';
+import * as welshTranslations from '../../../../../../assets/i18n/cy.json';
+import { createTranslateLoader } from '../../../../../../app/app.module';
 import { Http } from '@angular/http';
 
 describe('ProvidedEmailComponent', () => {
