+++ src/pages/communication-page/cat-b/components/provided-email/provided-email.ts
